Feat: 가격 int -> long 으로 변경

+++ src/main/java/site/keydeuk/store/domain/coupon/dto/response/CouponResponse.java
@@ -12,9 +12,9 @@ public record CouponResponse(
         @Schema(description = "쿠폰 이름", example = "깜짝 쿠폰")
         String name,
         @Schema(description = "쿠폰 가격", example = "4000")
-        Integer price,
+        Long price,
         @Schema(description = "주문 최소 금액", example = "40000")
-        Integer minPrice,
+        Long minPrice,
         @Schema(description = "쿠폰 만료 기간", example = "2024-07-12T00:00:00")
         LocalDateTime expiredAt
 ) {
