+++ src/main/java/site/keydeuk/store/domain/likes/dto/response/LikedProductsResponse.java
@@ -7,6 +7,6 @@ public record LikedProductsResponse(
         Integer productId,
         String productImg,
         String productName,
-        Integer price
+        Long price
 ) {
 }
